Scripts/requirements.txt
absl-py==0.11.0
alabaster==0.7.12
asn1crypto==1.4.0
astroid==2.1.0
astropy==4.0
atomicwrites==1.2.1
attrs==18.2.0
azure-common==1.1.26
azure-core==1.12.0
azure-nspkg==2.0.0
azure-storage-blob==12.8.0
azure-storage-common==0.37.1
azure-storage-nspkg==3.1.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
bleach==3.0.2
bokeh==1.0.2
boto==2.49.0
boto3==1.9.253
botocore==1.12.253
Bottleneck==1.2.1
category-encoders==2.2.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
ci-info==0.2.0
Click==7.0
cloudpickle==0.6.1
cmdstanpy==0.9.5
colorama==0.4.1
combo==0.1.1
comtypes==1.1.7
configparser==5.0.0
confuse==1.0.0
contextlib2==0.5.5
convertdate==2.2.2
cryptography==3.3.1
cycler==0.10.0
Cython==0.29.2
cytoolz==0.9.0.1
dask==1.0.0
decorator==4.3.0
defusedxml==0.5.0
distributed==1.25.1
docutils==0.14
elementpath==2.0.5
emailvalidator==0.3
entrypoints==0.2.3
ephem==3.7.7.1
et-xmlfile==1.0.1
etelemetry==0.2.2
ez-setup==0.9
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
Flask-Cors==3.0.7
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
future==0.17.1
futures==3.1.1
gevent==1.3.7
glob2==0.6
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.34.0
h5py==2.8.0
hdbcli==2.5.101
heapdict==1.0.0
holidays==0.10.3
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.18.1
idna==2.8
ijson==3.1.4
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.6
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
jmespath==0.10.0
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-server==0.2.0
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==17.0.0
kiwisolver==1.0.1
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.26.0
locket==0.2.0
LunarCalendar==0.0.9
lxml==4.5.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==4.3.0
mpmath==1.1.0
msgpack==0.5.6
msrest==0.6.19
multipledispatch==0.6.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nibabel==3.1.1
nltk==3.4
nose==1.3.7
notebook==5.7.4
numba==0.41.0
numexpr==2.6.8
numpy==1.19.4
numpydoc==0.8.0
oauthlib==3.1.0
olefile==0.46
openpyxl==2.5.12
opt-einsum==3.3.0
oscrypto==1.2.1
packaging==18.0
pandas==1.1.2
pandocfilters==1.4.2
param==1.10.1
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
petl==1.6.8
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.14.0
prov==1.5.3
psutil==5.4.8
py==1.7.0
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.9
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.2.2
PyMeeus==0.3.7
pymongo==3.11.0
pyod==0.8.3
pyodbc==4.0.25
pyOpenSSL==18.0.0
pyparsing==2.3.0
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
rdflib==5.0.0
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.11.0
setuptools-git==1.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==1.2.1
snowflake-connector-python==2.4.1
snowflake-sqlalchemy==1.2.4
sortedcollections==1.0.1
sortedcontainers==2.1.0
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.2.15
statsmodels==0.9.0
suod==0.0.4
sympy==1.3
tables==3.4.4
tblib==1.3.2
tensorboard==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.3
thrift==0.13.0
tools==0.1.9
toolz==0.9.0
tornado==5.1.1
tqdm==4.40.2
traitlets==4.3.2
twilio==6.44.2
unicodecsv==0.14.1
urllib3==1.25.11
validate-email-address==1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
WTForms==2.3.3
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
xmltodict==0.12.0
zict==0.1.3
